Uploaded Data Files
Use the Reformated-Data files to run the R code provided here. The original data files are available here as well for reference. The only difference between these files is that in the Reformated-Data files, all of the information is averaged and sorted for simpler data analysis. If use of the original data is desired, the variable names in the R markdown should be fairly simple to understand to receive the same results.